backend/pyproject.toml
[project]
name = "content-hub-backend"
version = "0.1.0"
description = "Personal feed aggregator that collects content from RSS feeds and web sources."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "feedparser>=6.0.12",
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
